@@ -1,13 +1,12 @@
 module github.com/jer8me/CertStore
 
-go 1.20
+go 1.21
 
 require (
 	github.com/mattn/go-sqlite3 v1.14.17
 	github.com/spf13/cobra v1.7.0
 	github.com/stretchr/testify v1.8.4
 	golang.org/x/crypto v0.12.0
-	golang.org/x/exp v0.0.0-20230811145659-89c5cff77bcb
 )
 
 require (
