@@ -8,8 +8,8 @@ import (
 	"github.com/jer8me/CertStore/pkg/store"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"golang.org/x/exp/slices"
 	"path"
+	"slices"
 	"strings"
 	"testing"
 )
